feat: move open actions to publications
