+++ apps/web/src/components/Publication/OpenActions/Nft/Mint/Metadata.tsx
