+++ apps/web/src/components/Publication/OpenActions/Nft/Mint/Price.tsx
