+++ apps/web/src/components/Publication/OpenActions/Nft/Mint/index.tsx
