+++ apps/web/src/components/Publication/OpenActions/Snapshot/VoteProposal.tsx
